Adjusted the value of dpiJsonNode.nativeTypeNum to be DPI_NATIVE_TYPE_FLOAT instead of DPI_NATIVE_TYPE_DOUBLE (#174).

+++ test/test_4300_json.c
@@ -1163,7 +1163,7 @@ int dpiTest_4311_bindJsonArrayFloatValues(dpiTestCase *testCase,
     array = &node->value->asJsonArray;
     for (i = 0; i < array->numElements; i++) {
         node = &array->elements[i];
-        if (dpiTestCase_expectDoubleEqual(testCase, node->value->asDouble,
+        if (dpiTestCase_expectDoubleEqual(testCase, node->value->asFloat,
                 valueToPass[i]) < 0)
             return DPI_FAILURE;
     }
